Big Changes
Completely revamped the internal design of my package page (external appearance remains unchanged).  It now pulls from the packages section of the DB and uses those to populate the page contents (an imgsrc field needed to be added to the db to make it work, so bear in mind it'll be broken on anyone elses computer until that field gets added in). Next up is making the actual booking page for the packages